requirements.txt
absl-py==1.0.0
appnope==0.1.3
arrow==1.2.2
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
blis==0.7.7
cachetools==5.0.0
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
codecarbon==2.1.3
cycler==0.11.0
cymem==2.0.6
Cython==0.29.30
DAWG-Python==0.7.2
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.3.0/de_core_news_sm-3.3.0-py3-none-any.whl
decorator==5.1.1
docopt==0.6.2
el-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/el_core_news_sm-3.3.0/el_core_news_sm-3.3.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.3.0/es_core_news_sm-3.3.0-py3-none-any.whl
et-xmlfile==1.1.0
executing==0.8.3
fi-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fi_core_news_sm-3.3.0/fi_core_news_sm-3.3.0-py3-none-any.whl
filelock==3.6.0
flatbuffers==2.0
fonttools==4.33.3
ftfy==6.1.1
fuzzywuzzy==0.18.0
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.0
h5py==3.6.0
huggingface-hub==0.5.1
idna==3.3
importlib-metadata==4.11.3
ipdb==0.13.9
ipython==8.3.0
it-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/it_core_news_sm-3.3.0/it_core_news_sm-3.3.0-py3-none-any.whl
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
ko-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ko_core_news_sm-3.3.0/ko_core_news_sm-3.3.0-py3-none-any.whl
langcodes==3.3.0
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
msgpack==1.0.3
murmurhash==1.0.7
nb-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/nb_core_news_sm-3.3.0/nb_core_news_sm-3.3.0-py3-none-any.whl
nl-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/nl_core_news_sm-3.3.0/nl_core_news_sm-3.3.0-py3-none-any.whl
numpy==1.22.3
oauthlib==3.2.0
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
preshed==3.0.6
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==8.0.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
Pygments==2.12.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pynvml==11.4.1
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.3.0/ru_core_news_sm-3.3.0-py3-none-any.whl
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
sentencepiece==0.1.97
six==1.15.0
sklearn==0.0
smart-open==5.2.1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.2.0
termcolor==1.1.0
thinc==8.0.15
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0
tqdm==4.64.0
traitlets==5.1.1
transformers==4.18.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
Werkzeug==2.1.2
wordfreq==3.0.1
wrapt==1.14.1
zipp==3.8.0
